backend/requirements.txt
django>=5.2,<5.3
djangorestframework>=3.16,<4.0
drf-spectacular>=0.28,<1.0
django-cors-headers>=4.6,<5.0
pytest>=8.4,<9.0
pytest-django>=4.11,<5.0
dj-database-url>=2.2,<3.0
